refactor(experiment-tag): simplify the consent gate wiring
Three reductions, no behavior change:

- Drop the buffer clears in setConsentStatus and releaseGate. Events only
  enter the buffer while no deferral is active, and the stash-time clear
  wipes anything earlier, so both cleared a provably empty buffer. With
  the hadDeferral clear gone, releaseGate is two lines — inlined at both
  call sites and deleted.
- Move the denial-cleanup arming block from initialize() into
  armDenialCleanup in consent-gate.ts, next to the cleanupArmedManager
  field whose invariant it maintains.
- Take a plain optional instanceName in clearAllPersistedData and
  clearIfErasedElsewhere instead of a one-field options object, matching
  getPersistedDataKeys.
