+++ packages/experiment-tag/src/consent/consent-gate.ts
@@ -1,5 +1,6 @@
 import { ConsentStatus, InitConfigs, WebExperimentConfig } from '../types';
 
+import { clearAllPersistedData, markIdentityErased } from './clear-data';
 import { ConsentManager } from './consent-manager';
 
 /** Returns a known status, or `null` if the value is not recognized. */
@@ -47,11 +48,44 @@ export const consentGate: ConsentGate = {
   cleanupArmedManager: null,
   reset() {
     // Replacing the manager detaches its listeners with it; the armed-manager
-    // comparison in `initialize` re-arms against the replacement on its own,
-    // so the null here is fresh state, not a correctness requirement.
+    // comparison in `armDenialCleanup` re-arms against the replacement on its
+    // own, so the null here is fresh state, not a correctness requirement.
     this.manager = new ConsentManager();
     this.deferredStart = null;
     this.started = false;
     this.cleanupArmedManager = null;
   },
+};
+
+/**
+ * Arms the denial cleanup against the current manager, once. Called from
+ * `initialize` because the sweep needs the apiKey that call supplies.
+ *
+ * Ordering matters: the immediate sweep covers a denial that resolved before
+ * this point (config value, or a setConsentStatus call against the pre-init
+ * stub), and the listener registered after it covers every later revocation.
+ * Registering second keeps a single sweep per denial rather than double-firing
+ * on the transition that just happened.
+ */
+export const armDenialCleanup = (
+  apiKey: string,
+  instanceName?: string,
+): void => {
+  if (consentGate.cleanupArmedManager === consentGate.manager) {
+    return;
+  }
+  consentGate.cleanupArmedManager = consentGate.manager;
+  const clearData = () => {
+    clearAllPersistedData(apiKey, instanceName);
+    // The sweep is origin-local; the marker is what crosses subdomains.
+    markIdentityErased(apiKey);
+  };
+  if (consentGate.manager.getStatus() === 'denied') {
+    clearData();
+  }
+  consentGate.manager.onChange((status) => {
+    if (status === 'denied') {
+      clearData();
+    }
+  });
 };
